+++ packages/be-ts-db-lib/src/repositories/BaseRepository.spec.ts
@@ -24,7 +24,7 @@ it('can create a new document and delete it', async () => {
 
 it.skip('can add a new test user', async () => {
   const userRepository = UserRepository.getRepo();
-  const newUser = UserSchema.parse({ userName: 'demoUser2' });
+  const newUser = UserSchema.parse({ userName: 'workoutIntegrationTestUser' });
   newUser.auth.password = crypto.randomUUID();
   const createResult = await userRepository.insertNew(newUser);
   expect(createResult).toBeTruthy();