+++ packages/be-ts-lib/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 🔖 [3.0.12] (2026-02-07)
+
+### 🏗️ Changed
+
+- Updated dependency: now requires `@aneuhold/core-ts-api-lib@^3.0.12`.
+
 ## 🔖 [3.0.11] (2026-02-06)
 
 ### 🏗️ Changed
@@ -162,6 +168,7 @@ No direct code changes; version bump for compatibility with new major versions o
 
 <!-- Link References -->
 
+[3.0.12]: https://github.com/aneuhold/ts-libs/compare/be-ts-lib-v3.0.11...be-ts-lib-v3.0.12
 [3.0.11]: https://github.com/aneuhold/ts-libs/compare/be-ts-lib-v3.0.10...be-ts-lib-v3.0.11
 [3.0.10]: https://github.com/aneuhold/ts-libs/compare/be-ts-lib-v3.0.9...be-ts-lib-v3.0.10
 [3.0.9]: https://github.com/aneuhold/ts-libs/compare/be-ts-lib-v3.0.8...be-ts-lib-v3.0.9